test(journal): split JournalTest
